refactor(#44): 매물 찾기 서비스 패키지 변경
